refactor: Changes the declaration of some functions.

+++ ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: arabelo- <arabelo-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/09 16:58:15 by arabelo-          #+#    #+#             */
-/*   Updated: 2023/04/09 17:59:09 by arabelo-         ###   ########.fr       */
+/*   Updated: 2023/04/11 21:23:37 by arabelo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, size_t size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
 	size_t	dest_len;
 	size_t	src_len;
